% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{malign}
\alias{malign}
\alias{malign-package}
\alias{insult}
\alias{dis}
\alias{snark}
\alias{zing}
\title{Positive encouragement is nifty, but have ya tried \emph{negative encouragement}?}
\usage{
malign(template = "You are ${adjective}!")

insult(template = "You are ${adjective}!")

dis(template = "You are ${adjective}!")

snark(template = "You are ${adjective}!")

zing(template = "You are ${adjective}!")
}
\arguments{
\item{template}{Character scalar, the template string.}
}
\description{
Everyone needs to be taken down a peg once in a while, especially
oneself! Forked from the complementary (and complimentary!) package praise,
malign provides ready-made insults for use when code does run not as expected.
Useful for troubleshooting code, or just to make oneself feel worse.
}
\details{
Replace parts of the template with random words from the malign word lists.
See examples below.
}
\examples{
malign()

## Capitalization
malign("${Exclamation}! This ${rpackage} is ${adjective}!")

## All upper case
malign("${EXCLAMATION}! You have done this ${adverb_manner}!")
}
\author{
\strong{Maintainer}: Joseph Brown \email{phylo.jwb@gmail.com} (\href{https://orcid.org/0000-0002-3835-8062}{ORCID})

}
