% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{malign_parts}
\alias{malign_parts}
\title{Parts of speech for maligning}
\format{
Named list of character vectors. List entries: \describe{
  \item{adjective}{Words and phrases to be used as negative adjectives.}
  \item{adverb}{Adverbs.}
  \item{adverb_manner}{Adverbs of manner, with negative meanings.}
  \item{created}{Negative synonyms of \sQuote{create} in paste tense.}
  \item{creating}{Negative synonyms of \sQuote{create}, in present participle form.}
  \item{exclamation}{Negative exclamations.}
  \item{rpackage}{Synonyms for the term \sQuote{R package}.}
  \item{result}{Words and phrases used to connote negative results.}
}
}
\usage{
malign_parts
}
\description{
Parts of speech for maligning
}
\keyword{datasets}
